finserv/requirements.txt
aniso8601==1.3.0
click==6.7
Flask==0.12.2
flask-restplus==0.10.1
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
kafka-python==1.3.4
MarkupSafe==1.0
python-dateutil==2.6.1
pytz==2017.2
six==1.10.0
Werkzeug==0.12.2
